hack for pow

+++ foundation/blockchain/genesis/genesis.go
@@ -11,7 +11,7 @@ type Genesis struct {
 	ChainID       uint16            `json:"chain_id"`
 	TransPerBlock uint16            `json:"trans_per_block"`
 	Difficulty    uint16            `json:"difficulty"`
-	MiningReward  uint16            `json:"mining_reward"`
+	MiningReward  uint64            `json:"mining_reward"`
 	GasPrice      uint16            `json:"gas_price"`
 	Balances      map[string]uint64 `json:"balances"`
 }
